Loss Ops: Make sigmoidCrossEntropyWithLogits private
