+++ src/ops/loss_ops.ts
@@ -24,7 +24,6 @@ import {assertShapesMatch} from '../util';
 import {expandShapeToKeepDim} from './axis_util';
 import {minimum} from './binary_ops';
 import {op} from './operation';
-import {sigmoidCrossEntropyWithLogits} from './sigmoid_cross_entropy';
 import {ones, scalar} from './tensor_ops';
 
 export enum Reduction {
@@ -253,6 +252,42 @@ function logLoss_<T extends Tensor, O extends Tensor>(
   return computeWeightedLoss(losses, $weights, reduction);
 }
 
+function sigmoidCrossEntropyWithLogits_<T extends Tensor, O extends Tensor>(
+    labels: T|TensorLike, logits: T|TensorLike): O {
+  const $labels =
+      convertToTensor(labels, 'labels', 'sigmoidCrossEntropyWithLogits');
+  const $logits =
+      convertToTensor(logits, 'logits', 'sigmoidCrossEntropyWithLogits');
+  assertShapesMatch(
+      $labels.shape, $logits.shape, 'Error in sigmoidCrossEntropyWithLogits: ');
+
+  /**
+   * Implementation Details:
+   *
+   * For brevity, let `x = logits`, `z = labels`.  The logistic loss is
+   *     z * -log(sigmoid(x)) + (1 - z) * -log(1 - sigmoid(x))
+   *   = z * -log(1 / (1 + exp(-x))) + (1 - z) * -log(exp(-x) / (1 + exp(-x)))
+   *   = z * log(1 + exp(-x)) + (1 - z) * (-log(exp(-x)) + log(1 + exp(-x)))
+   *   = z * log(1 + exp(-x)) + (1 - z) * (x + log(1 + exp(-x))
+   *   = (1 - z) * x + log(1 + exp(-x))
+   *   = x - x * z + log(1 + exp(-x))
+   *
+   *   For x < 0, to avoid overflow in exp(-x), we reformulate the above
+   *     x - x * z + log(1 + exp(-x))
+   *   = log(exp(x)) - x * z + log(1 + exp(-x))
+   *   = - x * z + log(1 + exp(x))
+   *
+   * Hence, to ensure stability and avoid overflow, the implementation uses
+   * this equivalent formulation:
+   *     max(x, 0) - x * z + log(1 + exp(-abs(x)))
+   */
+  const maxOutput = $logits.relu();
+  const outputXTarget = $logits.mul($labels);
+  const sigmoidOutput = $logits.abs().neg().exp().log1p();
+
+  return maxOutput.sub(outputXTarget).add(sigmoidOutput);
+}
+
 /**
  * Computes the sigmoid cross entropy loss between two tensors.
  *
@@ -295,7 +330,7 @@ function sigmoidCrossEntropy_<T extends Tensor, O extends Tensor>(
     $multiClassLabels = $multiClassLabels.mul(one.sub(labelSmoothingScalar))
                             .add(half.mul(labelSmoothingScalar));
   }
-  const losses = sigmoidCrossEntropyWithLogits($multiClassLabels, $logits);
+  const losses = sigmoidCrossEntropyWithLogits_($multiClassLabels, $logits);
 
   return computeWeightedLoss(losses, $weights, reduction);
 }
